Avoid running some Maven conversion tests on JDK8
In Maven conversion tests, if the Maven POM only specifies a source
or target compatibility, we only try to configure target compatibility
when the target compatibility does not match the source compatibility.

We assume JDK 1.8 when source or target compatibility is not specified,
which makes these tests fail for different reasons, depending on which
JDK we're using for the integration test.

In the future, we should be setting something else (like release), so
we don't have to make assumptions about the "default" source or target
compatibility.