Refactor the mapping system.
As the JUnit4 and JUnit5 processing was basically the same, it could be factored out into DefaultThrowableToTestFailureMapper, making a common abstraction.
I think we are on the correct level of abstraction here, as special cases will be handled by defining the list of  TestFailureMapper instances - and that is still up to the platform's implementation.