Avoid resolving artifacts when producing resolution result
The resolution result is calculated as part of the GRAPH_RESOLVED state of the configuration.
It is not necessary to reach ARTIFACTS_RESOLVED in order to have a full resolution result,
as the resolution result only describes the graph and not its artifacts.

This optimizes cases where ONLY the graph is resolved for a configuration, ie when the user
only wants to analyze the resolved graph without accessing its artifacts.