Do not store request details in local dependency metadata
These three fields were virtually unused. The attributes were unused, and the componentId and client configuration was only used for the
rare error message where the dependency declared a target configuration AND that target configuration was not found in
the target component. This case is rare, since it only occurs when variant-aware dependency resolution is disabled. And,
in no other places do we print the declared configuration. Furthermore, the componentId is already printed in error messages
where it was printed via the now-removed message, so no loss of information is apparent for the user.

This will likely result in a non-negligible memory savings.