docs: View and Debug Dependencies: add a note about `buildEnvironment`
Recently, I needed to look at a graph of dependencies from the
configuration `classpath` that is configured in the block `buildscript`.
This can be done using task `buildEnvironment`, which is described in
section "Reporting dependencies" of page "Command-Line Interface" in the
Gradle userguide.  Graph of dependencies from other configurations can
be rendered using task `dependencies`, which is described on the page
"View and Debug Dependencies".

Unfortunately, my search-engine-fu wasn't enough to find the description
of task `buildEnvironment`.  Instead I ended up on the page "View and
Debug Dependencies".

To help users like myself, add a pointer towards documentation of task
`buildEnvironment` to the section "List Project Dependencies" of the
page "View and Debug Dependencies".

Signed-off-by: Andrei Rybak <rybak.a.v@gmail.com>