Fail when attempting to reselect variant with explicitly declared artifact
The previous method would essentially allow reselection to occur as if the artifact was never requested, however this
can sometimes produce incorrect results as explicit artifacts can refer to any arbitrary artifact in the component. This means
the reselected artifact may not correlate to the originally explicitly requested artifact.

Instead, we update the behavior to fail. Users can avoid this by using a lenient artifactView