Ignore test when using config cache
This test explicitly checks that intra-project task parallelism s constrained again after a task finishes waiting on work items in between task actions.  With config cache, project mutability is not a concern and therefore intra-project task parallelism is allowed, so this test is not meaningful when config cache is enabled.