[WIP] Fail instrumentation if MRJAR contain non-instrumentable entries for...
...the current JVM to load.

This may happen if Gradle is running on a new, unsupported JVM. For
example, let's assume that Gradle can instrument bytecodes for up to
Java 21, and the MR-JAR supports Java 21 in the root directory and
contains versioned directory for Java 23. Running Gradle on Java 21 and
22 with this JAR on the class path is fine as the JVM doesn't load
bytecodes for Java 23. However, running Gradle on Java 23+ means that
these bytecodes are going to be loaded.