Configure StartParameter coming from TestLauncher on the client side
Configuring it before running the test action is too late when the
configuration cache is in play. CC uses StartParameter earlier to
calculate the cache key and uses original StartParameter for that. This
means that all TestLauncher invocations used the same default tasks for
their CC key, reusing the cache entry even when different tasks are
requested.