Move Gradle Enterprise plugin services to correct service scope
The problematic service was DefaultGradleEnterprisePluginServiceRef. This is intended to be build-tree scoped as per its annotation. However, it was being registered in Gradle scope. The net effect of this was that references to this object in objects across build boundaries were ending up with different instances upon deserialization. This broke the cross build scope access mechanism used by the Gradle Enterprise plugin.