Replace LGPL licensed Checkstyle XSLT with Apache licensed version
This change replaces the XSLT supplied with Gradle with one created
by me under the Apache license.

The existing XSLT was identified as an Apache licensed file when it
was checked into the Gradle codebase, but the origin of this file
is from the a LGPL licensed Checkstyle contribution repository.

The report retains the same features:
- A summary
- A clickable file listing sorted by total error count
- A list of violations for each file

The styling was changed to mimic the Gradle documentation styling.

This retains the same file name on the off chance that something
outside of Gradle relies on that name.