Add indirection layer when adapting intercepted call arguments
We were using `instanceof` directly but Groovy's rules for type coercion
are trickier. In order to support them, we have to call some non-trivial
logic when deciding if the actual argument value matches the expected
type.

This commit only introduces the indirection but the rules are still the
same. Actual type coercion will follow in the next commits.