Only emit planned transform operation in project context
It is possible to register transforms in the context of a script,
and those transforms might run on substituted included build dependencies.

Even though these transforms will never be part of any execution plan,
calculating the node identity of them will currently fail, because
they don't have an owning project.

This causes the Build Scan plugin to fail.

Therefore, we stop emitting the build operation in this case.