Make BuildLauncher and TestLauncher behavior more uniform
The TestLauncher calls AbstractTestTask.setIgnoreFailures(true) which
makes the build pass even if the test failed. This is an unexpected
behavior compared to the BuildLauncher as users do expect build failures.

The failures were ignored in order not to stop the execution when a test
task failed. At the same time, the client can always run the build with
the --continue flag and have the same behavior.