Cache default PgpSignatory when in-memory signing is used
Computing a PGPSecretKey from armored text representation is suprisingly
expensive. When in-memory PGP key is used, `getDefaultSignatory` gets
called a lot when deducing task dependencies, fingerprinting task inputs,
and storing the configuration cache, so caching is a good improvement.

This commit improves signing performance of the reproducer project with 5
signing tasks and about a dozen artifacts from 68s to 5s.