DefaultGradleEnterprisePluginAdapter is not a service
The adapter carries state and is referenced by the manager. It is
not a service in the sense that it can be re-created in a stateless
way when firing up the remaining services.

The configuration cache would not use serialization to restore a service,
but use the existing service since it assumes the service is stateless.