Iterate through attribute keys added by attributeProvider in insertion order
When iterating through the Attribute keys, we need to ensure we follow insertion
order.

"Lazy attributes" are attributes with a value defined by a Provider.

This should not typically change the result of resolution because all attributes
are eventually considered, but this was an underlying source of non-deterministic
behavior in #26298.