Allow attribute values to cause the realization of other attribute values
When calculating the value of one attribute, nothing prevents dependency
resolution or something else to look at other attributes in the same
attribute container.

When this happens, it's no longer safe to assume the keys obtained at the
start of realization are still valid when looking up the value of a lazy
attribute.

In this scenario, we can ignore the missing (already realized) attribute.