De-duplicate results for resolved variant dependencies in resolution result
At some point, this used a MultiMap with HashSet values. This caused order
of elements out of getDependenciesForVariant to be non-deterministic.

The MultiMap was switched to use ArrayList values. This fixed the inconsistent
ordering problem, but introduced a new problem.

With Set-like values, duplicate dependency results were filtered out.
With List-like values, duplicate dependency results are kept.

This commit changes the MultiMap's values to a LinkedHashSet. This
will have a stable (insertion) ordering and de-duplicate dependency
results.

fixes #26334