Extract java-base plugin from java plugin into separate project
This splits java-base and java plugins into their own subprojects. It also migrates some code and tests to more appropriate subprojects within the JVM platform.

This involves extracting the consistent resolution code from the plugin and putting it into the component.  Consistent resolution is now done for every JvmSoftwareComponentInternal (there should only be one of these).

In addition, this change consolidates the logic between the Java and Groovy plugins for configuring compile task conventions into a helper and makes it the responsibility of the groovy-base plugin.