Merge pull request #26411 Add build-level retry
### Context

We've been annoyed by non-test flakiness for a long time. This PRs adds the build level retry support for CI builds. For functional tests, if `GRADLE_RUNNER` step fails, there will be 3 extra steps running:

- KILL_ALL_GRADLE_PROCESSES: kill all Gradle JVMs on the build agent, including all daemons.
- CLEAN_UP_GIT_UNTRACKED_FILES_AND_DIRECTORIES: remove all on tracked files in the checkout directory.
- GRADLE_RETRY_RUNNER: rerun the failed builds.

In short, we retry the build step in a fresh checkout directory with a fresh environment. We hope this could fix most flakiness on CI. For now, this is only enabled for pre-tested commit builds.

Co-authored-by: Bo Zhang <bo@gradle.com>