Consolidate dependency graph visitors
The primary goal of this change is to "bring up" the artifact visiting logic into the DefaultConfigurationResolver so that
it can reside with the other graph visiting logic. Now, all logic that traverses constructed dependency graphs is located
in one place. This will allow us to start merging the concepts of graph visitors and artifact visitors, so that we can in the
future work towards adding more artifact information to resolution results -- or at least introduce some improved connections
between a resolved graphs and the artifacts it produces.

As a side effect, some work was done to reduce reliance on `ResolveContext` in this area
Additionally, DefaultArtifactDependencyResovler was renamed to DefaultResolverFactory to better reflect its behavior
Further, javadoc was added in several places that otherwise lacked any description.