Deprecate SelfResolvingDependency
In Gradle 9.0, ProjectDependency and FileCollectionDependency will no longer implement this interface

A ProjectDependency is a SelfResolvingDependency, but in practice it is not resolved that way by the dependency resolution engine.
In order to support the self-resolving methods, a ProjectDependency needs an actual Project instance (or an instance to the project registry)
This should not be necessary. A project dependency should be a simple reference to the identifier of a project and no more.

A FileCollectionDependency already exposes its FileCollection, so existing calls to the resolve method are easily replaced.

SelfResolvingDependencyInternal is still used by DefaultFileCollectionDependency, but cannot be renamed since KMP uses that internal interface to implement NPM dependencies.