Introduce `BuildLogicBuildQueue` service
To ensure only a single thread of a single daemon is ever running a particular
build logic build and reduce the window for build logic jars to be corrupted by
overlapping builds.

The initial implementation uses a reentrant lock and a file lock to ensure only
a single thread is ever running a build logic build from a particular root
build.

This strategy doesn't prevent different root builds that share the same included
builds from overlapping and possibly corrupting the resulting jar files but that
can be addressed in a future commit.