Ensure role-locked configurations are non-visible
Configurations with visible=true are automatically built when the assemble task is executed.
We do not want this behavior going forwards, so we disable visible on these new configuration types

Further add tests to verify current functionality where the assmeble task executes other tasks by default
Change behavior so that platform plugin variants are non-visible. Artifacts should not be declared on these configurations