Add plain text progress reporting
Introduces a PlainBuildStatusRenderer which provides progress reporting
to PlainProgress consoles. Progress reports are similar to the
RichBuildStatusRenderer, except they only include the build phase,
percentage, failure status, and terminal control characters are not
used.

To avoid flooding the output, progress is only published when one of
those three properties have changed.

Before this, the only way to see current build status was to use the
rich console, which is often impossible or undesirable for CI builds.

Progress updates look like:

    > Progress: CONFIGURING 0%
    > Progress: CONFIGURING 50%
    > Progress: EXECUTING 12%

Signed-off-by: Kevin Mark <kmark@palantir.com>