Remove outdated smoke test
This smoke test checked that AGP+KMP produced the expected metadata
when publishing.

The version of AGP and KGP are both old and it doesn't make sense for
our tests to verify that AGP/KGP is producing equivalent metadata over time.

We have other tests that exercise AGP, KGP and KMP separately with dependency
resolution.

This test _might_ catch problems where Gradle produces the "wrong" metadata,
but this isn't what we're testing for and it could be just as easy to accept
changes in metadata as "correct" behavior.