Assume unresolved packaging variables in a POM mean jar
Maven ignores the declared packaging and assumes jar unless told
otherwise.

Gradle attempts to resolve a dependency's artifact based on its packaging.

In the past, we would have attempted to resolve an artifact with the
extension ${packaging.type} and, failing that, tried jar.

Now, when we detect that the packaging isn't something valid, we
immediately default to jar in our internal representation of the POM.