Move most kotlin-dsl plugins integration tests out of k-d-p project
Kotlin DSL plugins depend on the Kotlin Gradle plugin (KGP), and because
of that the latter becomes a part of the test classpath. In order to
support property upgrades, KGP has to be bytecode-processed. However,
when embedded runner is used, KGP is loaded from the test classpath and
bytecode processing isn't applied.

To work around that, we move all plugin integration tests into a
separate subproject; it is easier than removing KGP dependency from the
k-d-p's integTestImplementation.