Improve performance of RelativePath canonicalization
This is a follow-up to https://github.com/gradle/gradle/pull/24943.

There was some uncertainty as to whether that change resulted in
a detectable drop in performance in Gradle's overall performance
tests. I put together [some benchmarks](https://github.com/AlexLandau/gradle-relative-path-perf)
looking at the constructor method that was modified, and based on
my probably-non-representative sample of test cases, the new
implementation took around 3.4x as long as the one that did not
perform canonicalization.

The version in this PR skips running canonicalization checks on
segments that come from existing `RelativePath` objects, which have
already been canonicalized. This implementation actually performs
better than the original, pre-canonicalization implementation in
my benchmark test; I would take this with a grain of salt, as I
do not have experience with JMH and may have used a poor setup
or poor choice of test cases. (There's at least some evidence
that for the range of test cases attempted, copying into the array
with a for loop is faster than `System.arraycopy`, which would
explain this. This could, of course, vary across JVM versions and
other factors.)

Signed-off-by: Alex Landau <alandau@cs.stanford.edu>