Create root platform subprojects
Create infrastructure to declare subprojects for individual platforms.
These platform projects are meant to aggregate tasks across all of their child projects.

For example, running ':jvm:test' will run all unit tests for the jvm platform subprojects.
This currently only works for tests, as we do not have the proper test result variants exposed for integ tests.
This will require us to migrate to test suites for our integ tests.

This has been enabled for the jvm and software platforms.
We cannot enable this across all platforms, as some plaforms contain projects with the same name as their platform.

For example, the root project for the jvm platform has the name ':jvm'.
The 'ide' platform already contains an ':ide' subproject. Same with 'enterprise'.
These will need to be renamed if they want to support platform aggregations