Do not relocate symlinked snapshots
If a snapshot was taken of symlinked content, we don't keep track of enough information about the link to tell if after relocation the snapshot would stay correct. To avoid problems, we refuse to relocate such snapshots.

This can be improved later by retaining symlinked snapshots where possible (e.g. when the symlink points within the relocated hierarchy), or re-take them already. This is not a goal now, though.