Make plugin version optional in version catalogs
In some cases, one can apply a plugin in `plugin { ... }` block without
specifying a version.
However, at the moment it's not possible to leave version unspecified in a
version catalog and then apply via `alias`. Under the hood `alias` calls
`version` unconditionally and then fails on null or empty value.
This commit prevents such a failure.

Use case for not specifying plugin versions in version catalog:
* a plugin may be applied from an included build (incl. `buildSrc`)
* plugin version may be pinned via settings, a settings plugin or an init
  script via `pluginManagement`
* a plugin may be discovered in build script classpath
* other situations and mechanisms unknown to me

Even if version is not specified, version catalog can be used as a bill of
materials or a reference list. Besides, generated DSL accessors work as
a typo prevention mechanism.

It happens that interface `PluginDependenciesSpec` already contains a
satisfactory implementation for the offending methods, so this is a
delete-only commit.

Signed-off-by: Valentin Kulesh <leshy@unshare.net>