Add update operation for Properties and ConfigurableFileCollection
The new operation provides a way to update the property/file collection
based on its current value without introducing a circular reference or
materializing the value.

This is mostly useful for the migrated eager properties where build
scripts may adjust the property values set by plugins. It is less
helpful for plugin authors because of the temporal coupling: the value
of the property at the moment update is called is not yet a final one,
and may be overriden.