Wire dependency resolution for dependency mapping
Version mapping and dependency mapping previously did not properly wire dependency resolution.
Instead, they resolved configurations eagerly instead of through providers.

This properly uses providers to wire resolution results all the way down to the dependency mapping resolvers.

This does not yet actually declare resolution properly, as that is blocked by #27255

This change primarily relies on a new provider, MergeProvider, which can accept a list of providers and provides a list of the resolved provider values