Deprecate publishing manven dependency with differing artifact name
Currently, when publishing to maven, Gradle interprets an artifact with an explicit name as a dependency on a module name equal to that artifact name.
This is an incorrect interpretation, as this changes the original declared coordinates.

This ability to declare artifacts with name different than the module's name is a feature only supported by Ivy.
This should always have been an error.