Decouple configuration repository filtering
Allow configurations to themselves choose which repositories to use instead of having the repositories know
which configurations to allow. The DSL is still configured on the repository itself, but it is the configuration's
reponsibility to filter repositories based on those options.

The biggest advantage is that repositories no longer need to know the name of the configuration being resolved