Revert changes that run dependency actions when calling getAllDependencies
This is to avoid potential surprising changes in behavior. Instead, we explicitly call it in hasDependencies to ensure the
short circuit empty configuration resolver does not accidentally skip over configurations that only have dependenices
defined with withDependencies

We also add dependencies from the collector to the configuration more eagerly so they are visible on the configuration's
getDependencies and getAllDependencies. Additionally, we add mutation guards to the dependencies in the collector and finalize
the collector's list of dependencies on read in order to avoid dependencies getting added or mutated after the set has been observed