Resolve classpath configuration only once
configuration. The second phase analyzed that configuration, and if the auto-applied plugins were
not present there, it would modify that configuration's dependencies and resolve it again with
new dependencies. The only reason this works is because we specially treat the classpath configuration
and clear all resolution state after it is resovled. This double-resolution with different dependencies
is very unconventional for a configuration so we should avoid it here.

This commit updates the two-phase resolution into a single phase. Instead of resolving, checking for
existence on the classpath, and resolving again, we resolve a single time. However, we place the
auto-applied plugin on the original resolution by adding a no-version dependency and a preferred-version
dependency constraint. This ensures any required version constraints from either explicit or transitive
plugin requests override the preferred version from the auto applied plugin

Furthermore, we improve the way alternative plugin coordinates are implemented -- specifically to handle
the transition from the Gradle Enterprise coordinates to the Develocity coordinates. Previously, during
the second resolution phase, we would not auto-apply plugins if they were present on the first phase classpath.
This however broke-down if the plugin was on the classpath but was not applied.

Instead, we enhance the plugin manager to be alternate-id aware, and also ensure we register the alternate
coordinates with the resolution so that conflict resolution can properly ensure two versions of the plugin
with different coordinates are not present on the classpath simultaneously.