Separate DefaultLenientConfiguration from ResolveContext
In attempt to reduce the places where ResolveContext is used, we update DefaultLenientConfiguration to only accept a
ResolutionHost -- a much weaker counterpart to ResolveContext. To do so, we change the description of the ResolveArtifactsDetails
to only use the ResolutionHost display name instead of the configuration path, and deprecate getConfigurationPath method on
the build operation. Te deprecated method on this build operation was confirmed to not be in use by Develocity in any supported version.

This allows us to remove several methods from ResolveContext