Deprecate ResolvedConfiguration without nagging
For quite a long time we have considered ResolvedConfiguration, LenientConfiguration, and its associted types
as legacy but have neglected to inform the community of our opinion. We have hesitated for some time since adding
deprecation warnings can be quite noisy, especially when the APIs being deprecated are used by many existing builds
and plugins.

However, despite the existing usage of this API across the community, this API has been superseded for quite some time
and as such we should begin the process of moving away from it. The newer ResolutionResult and ArtifactCollection
APIs are more powerful, and are actively maintained. ResolvedConfiguration and its related APIs have been in maintence
mode for some time.

It is important that we remove APIs like this -- those that have newer, feature-complete replacements -- as one common
complaint about Gradle is that there are too many ways to do the same thing. We are listening to the community here and
removing one of the superfluous ways.

We are simply adding deprecation annotations at this point and hope that their presence can nudge users towards moving
away from these APIs. As more builds and popular community plugins move away from this API, we will add proper deprecation
warnings.

The ResolutionResult API is scheduled for removal in Gradle 10.