Separate LenientConfiguration and VisitedArtifactSet
VisitedArtifactSet is a 'new' type and LenientConfiguration is an old public API.
Our DefaultLeneintConfiguration was our only implementation of VisitedArtifactSet.
This commit splits these implementations in to, so that we don't need to use a legacy
type in order to implement a 'new model' type.

Along the way, simplify DefaultConfiguration by allowing ResolutionResultProviders to be mapped
Add artifact sort order to ArtifactSelectionSpec, since choosing the order can be deferred to selection-time
Fix work-around for needing to cache exception thrown by artifactsAccessed in dependency verification override
Pull-out generic implementations of SelectedArtifactSet and VisitedArtifactSet so we don't need build-dependencies-only versions
Pull-out ResolvedArtifactSetResolver logic, allowing ResolvedArtifactSets to be easily visited. Before you needed a LenientConfiguration to do this.