Downgrade DependencyCollector assertion to warning
A recent assertion added to DependencyCollector is being hit by Kotlin projects.
Even though this is an incubating feature, we have decided to downgrade this to a warning
due to how easy it is to reproduce. This will become a warning in Gradle 9.0

Along the way, a bug was discovered relating to addAllLater and withType. Tests
have been added to reproduce this bug and have been annotated with @NotYetImplemented.

Finally, smoke tests have been added to verify Kotlin JVM and Test Suites work together.