Extract artifact resolution logic from DefaultLenientConfiguration
This commit moves some important artifact resolution logic out of DefaultLenientConfiguration.
Currently DefaultLenientConfiguration is our only fully capable VisitedArtifactSet implementation,
however it also implements many legacy public API methods, many of which accept a dependency spec
for filtering artifacts -- a filtering parameter we are trying to get away from.

There are advantages to separating implementations of public and internal APIs. Right now, it is
difficult to refactor artifact selection logic since it is so tightly coupled to DefaultLeneintConfiguration,
a type that cannot be significantly modified due to it implementing public APIs.

In this commit, we prepare for the separation of DefaultLenientConfiguration and VisitedArtifactSet. We
pull out a concrete implementation of SelectedArtifactSet and pull out logic that converts a ResolvedArtifactSet
into a SelectedArtifactSet.

Along the way, we simplify the implementation of DefaultVisitedArtifactResults, add a SortOrder parameter to
ArtifactSelectionSpec, implement a map method for ResolutionResultProvider, and make processing of ResolverResults
easier in DefaultConfiguration by using this map method.