Consolidate local component observation
The ResolvedLocalComponentsResultGraphVisitor collects local components in a resolved graph and produces a result containing
all resolved projects and their consumed configurations. Then, the resolvable configuration is responsible for notifying
the rest of the build that these projects were consumed.

This can be inaccurate in some edge cases, and we should instead move this to the LocalComponentRegistry, where we track
component observation at the time they are observed.

Furthermore, we use this mechanism to mark individual configurations as observed so they can no longer be mutated. This
is not comprehensive enough, and we will need to expand this into a more robust observation mechanism built directly into
DefaultLocalConfigurationMetadataBuilder.

This commit prepares us for these above changes by reducing the exposure of ResolvedLocalComponentsResultGraphVisitor
and its relevant types. We migrate as much logic into the visitor iself. This makes it much simpler to in the future
handle each of the two above use cases seprately while minimizing thier impact.