Improve instantiation of DependencyGraphBuilder
DependencyGraphBuilder was previously instantiated once for each resolution. Furthermore, it accepted
a ResolutionStrategyInternal in its 'resolve' method.

The initial goal of this commit was to remove the resolution strategy as a parameter, as this is a
public type, highly coupled to 'Configuration'. Now, we pass in only the relevant fields from the resolution strategy.

Given the number of parameters passed into the resolve method and constructor of the builder, it made
sense to separate them based on services and configurable parameters. Now, we make the builder build
scoped, and pass in all user-configurable values into the 'resolve' method.