Rework `UserInputHandler` to allow better interaction with the user.
The API now groups a set of questions into a single batch, or "interaction". This allows the UI to better deal with the user being asked a series of related question, for example in the presence of concurrency.

For now, the implementation simply avoids repeatedly redrawing and erasing the rich console status bars between related questions in order to get rid of some flicker and avoid the separator between questions being erased.

Later, the implementation might apply some concurrency control to avoid interleaving related question with other console output or other questions.