Track reasons for skipped tests
It is frustrating when for some reason you try to run tests and for some reason they are all skipped.
In this situation, Gradle does not tell you _why_ they were skipped.
Now it does.

In this commit, we track the reason for any given test being skipped.
At the end of the build, if no tests were selected, we add the skipped reasons to the message.
We also deprecate running tests where all tests are skipped, similar to how we have deprecated running tests where all tests are filtered.