Deprecate creating configurations in buildscript block
The intention of the buildscript block is to contain the 'classpath' configuration and allow dependencies to be declared on the configuration
This commit deprecates using the block for other questionable purposes, by deprecating creation of new configuration in the block.

This closes some odd corner cases where a buildscript configuration can resolve other buildscript configurations, a behavior that we never intended to support.
There are some questionable use-cases for using the buildscript block to perform adhoc resolutions, for example in settings scripts where there is no other way to obtain a configuration.
For this reason, creating detached configurations are still permitted in buildscript blocks.

We add a bunch of tests in BuildscriptResolutionIntegrationTest.groovy, which tests corner cases for buildscript configuration resolution.