Incremental Compilation - prefer project compiled sources (gradle/gradle#21255)
This stops a situation where if versions of the same class are in both your sources and your dependencies, the version in your dependencies would be used over the one in your sources. This would lead to compile errors in incremental builds in instances where a consumer class in your sources depends on the specifics of your sources' version that differs from the version in the dependencies.

Signed-off-by: Zed Spencer-Milnes <zed@ziax.com>