Manually fix the expected set of "symbols" in a Swift application or library
This isn't really correct, but it's closer to what we were asserting on.

We should re-evaluate what we mean by "debug symbols". It would probably
be easier to rely on certain de-mangled symbol names, since these should
be the same across all ecosystems and tools.

Currently, we use readelf to extract symbols from Swift binaries and
the list of files we get from readelf is different from C/C++ binaries.