Adjust redundant configuration usage change warning
Previously, we emitted a deprecation warning in some cases when a usage for a configuration was set to a value it already had.
This was only emitted in certain cases -- for non-legacy, non-detached configurations, and when the value-being-set was to enable the usage.

The goal of this warning is to push users to not try to set roles for configurations that already have their roles set upon creation.
However, the previous implementation is not ideal for the two following reasons

1. The warning message was worded in a way to dissuade users from redundantly changing a configuration's usage, making it seem this was a sort of style/linting deprecation as opposed to a correctness one
   All other cases of non-redundant changes to roles were handled with a different deprecation warning
2. We only warned about this redundant change in very narrow circumstances -- when the usage's value was being set to true and for configurations with certain names
   This effectively defeats the purpose of this deprecation. The name matching avoided our own smoke tests from emitting this warning, but there are scenarios that users were hitting where this name matching did not cover.

Really, the goal here is for build logic to not call setCanBeX methods on configurations where the role is already set upon creation.
For this reason, we update redundant calls to setCanBeX to emit the same warning as calling these methods non-redundantly.
However, since KMP still redundantly calls setCanBeX, we hide emitting of this deprecation behind a system property.
We plan to file an issue with KMP to stop calling setCanBeX on all configurations that already exist -- or those that are managed by Gradle / created via source sets -- and tell them to verify their changes by enabling this system property.
Until then, we will not warn for redundant usage changes.