Update supported CentOS version for file-system watching
In a weird migration story CentOS 8 (which we did support) has already been deprecated, while CentOS 7 is still supported (but we did _not_ support FSW on it). Now there's CentOS _Stream_ 8 (which we _do_ support, and have tests for it), though it's going to reach end-of-life at the end of May 2024. After that there will only be CentOS Stream 9.

Details: https://blog.centos.org/2023/04/end-dates-are-coming-for-centos-stream-8-and-centos-linux-7/