Speed up runtime classpath fingerprinting
The RuntimeClasspathResourceHasher does not inspect the contents of JAR entries,
but just hashes them in their entirety. This means we don't have to unpack them
and can instead hash their compressed bytes. This both saves the unpacking time
and means less bytes to be hashed.

The ZipFile API does not give direct access to the raw input stream, but it is
easy to obtain by unwrapping the InflaterInputStream.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>