Map versions of dependencies with artifacts
92ae9e639e604f57c65343fc7a2b97a94b511da5 changed behavior to not map dependencies with artifacts.
The intention was to keep the root coordinates when mapping dependencies with artifacts, since that is how Gradle resolution behaves.

This change reverts the above commit to go back to mapping dependencies with artifacts, but instead mapping to component-level precision.
For legacy version mapping, this is equivalent to restoring prior behavior, as legacy version mapping only maps to component-level precision.