Document resolution reliability reproducers
Consolidate several reported reproducers for resolution issues.
All of these reproducers are related to issues with reliability or correctness.

In some cases, resolution fails with a 'corrupted resolution result' failure, in some cases, capability conflits
are ignored, and in some cases dependencies are skipped altogether.

Many reproducers have been transformed from their original form to versions that can be easily executed
as a Gradle integration test.
Some reproducers still need to be made more minimal, and their originals have been kept around so that after
we resolve the minified version, we can ensure we resolved the original issue -- as the minimized reproducers
may be very different than the originals