Surface errors committing config cache report
Previously if the executor shutdown timed out, `lateinit var reportFile`
would still be uninitialised, which hides the actual underlying problem
(that there were presumably tasks on the executor that didn't complete).

By submitting a `Callable` and waiting on the resulting `Future<File>`
we will surface the actual problem (almost certainly a
`TimeoutException`) which will help in diagnosis.

Also, hopefully log the tasks that didn't complete which may also help
in diagnosis.

Signed-off-by: Robert Elliot <rob@lidalia.org.uk>