Handle missing extension when publishing and consuming GMM
In some cases, the extension field of a published artifact is missing when the type field is present.
This happens when using the 'artifact { classifier = 'foo' }} syntax.

When this happens, we publish the GMM without the extension field, and when consuming that dependency, we
look for an artifact without an extension. This differs from the behavior of local resolution, where we default
the extension to the type if there is no extension present.

This commit updates behavior to publish the extension field as the type field, if the extension field is missing,
thus new versions of Gradle will publish GMM that work for old broken versions of Gradle.
We also update behavior to default the extension field to the type field when consuming GMM with a missing extension
field, thus allowing new versions of Gradle to consume broken GMM published by old versions of Gradle.