Decouple dependency locking from configurations
Dependency locking was written assuming configurations were the only thing that could be resolved.
Shortly, this will no longer be true, so we update dependency locking to avoid this assumption.

In practice, this mostly means we introduce the concept of a 'LockId', which is the name attached
to the dependenices for a given resolution within a lockfile. For a configuration, the lockId is equal
to the configuration name.