Do not select root node for constraints
The root node is always selected.
Previously, constraints would add incoming edges to the root node if the constraint was on the same module as the root node
There is a deprecation warning to ensure no dependencies select the root node.
We avoid that deprecation warning for this case, since we do not need to add edges to the root node