Allow multiple test suites with the same test type
Test report aggregation and jacoco aggregation use the test suite type attribute to select a single variant in a remote
project that has that type. There is a restriction that only a single test suite in a project can have a given test type
otherwise the variant reselection used by aggregation would run into ambiguous variant errors.

However, implementing this restriction is very difficult, as the test suite type is a function of a lazy property, and thus
the validation step must be run late. Currently, this is run whenever any configuration in the project is resolved, but this
is a side-effect of inefficiently realizing configurations' properties. Eventually, this side effect will go away and we will
need to address the point in time where this validation occurs.

This commit addresses this by removing the restriction entrely.

TODO: talk about the fanout