Improve error reporting when repositories are disabled
This tracks the reason for a repository being disabled, and adds it as the cause of any subsequent errors that short-circuit to report only the fact that a repo was disabled.  Tracking the ultimate cause like this allows for always printing it.

The BuildExceptionReporter gains some additional intelligence to avoid repeating failure trees which stem from the same cause.

This change also:
- Adds some additional tests around scenarios where (possibly improper) manual exception handling would have previously hidden the original cause - it is now shown.
- Displays the name of the repo that has been disabled when displaying the failure.
- Renames blacklist to disabled list throughout.