Expose configurations as variants
A ComponentGraphResovleState has a getConfiguration method, however this is a legacy construct. Gradle should operate
only on variants. An integ test was added showing how using this method causes ivy to in some cases consume a non-consumable
configuration.

We add a new mechanism to get a variant of a component that is identified by a configuration name. This method is safe
to access configuration-identified variants, ensuring that the configuration is consumable. Further, it exposes
the varaint typed as as a variant, ensuring the returne variant cannot be used as a configuration.

There are still come cases where the older getConfiguration method is used. Especially in the ivy graph selection
code. In the future, we can work to remove this method.